requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
ansi-escapes==0.1.1
ansi-styles==0.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
autopep8==2.0.1
awkward==1.10.2
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorlog==6.7.0
ConfigUpdater==3.1.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dict==2020.12.3
dill==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
fastjsonschema==2.16.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.31
graphnet==0+untagged.2245.g3c87005
graphviz==0.20.1
greenlet==2.0.2
has-flag==0.1.1
hydra-core==1.3.1
icecube==0.0.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.11.5
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==7.4.9
jupyter-console==6.5.1
jupyter_core==4.12.0
jupyter-events==0.6.3
jupyter-server==1.23.6
jupyter_server_fileid==0.7.0
jupyter_server_ydoc==0.6.1
jupyter-ydoc==0.2.2
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
jupyterlab-widgets==3.0.5
kaggle==1.5.12
kiwisolver==1.4.4
lightning-utilities==0.6.0.post0
lion-pytorch==0.0.6
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.6
omegaconf==2.3.0
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
plotly==5.13.0
polars==0.16.6
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.22.0
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==11.0.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.0
pytorch-lightning==1.9.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.15.0
setproctitle==1.3.2
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.46
supports-color==0.1.2
tenacity==8.2.1
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
timer==0.2.2
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.1+cu116
torch-cluster==1.6.0+pt113cu116
torch-geometric==2.2.0
torch-scatter==2.1.0+pt113cu116
torch-sparse==0.6.16+pt113cu116
torch-spline-conv==1.2.1+pt113cu116
torchaudio==0.13.1+cu116
torchmetrics==0.11.1
torchvis==0.4.0
torchvision==0.14.1+cu116
torchviz==0.0.2
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typed-ast==1.5.4
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.14
values==2020.12.3
wandb==0.13.10
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
wheel==0.37.1
widgetsnbextension==4.0.5
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
zipp==3.13.0
